Start to fix bugs
